Euclidean

Euclidean beta diversity metric.

Usage

euclidean(counts, weighted = TRUE, pairs = NULL, cpus = n_cpus())

Arguments

counts

An OTU abundance matrix where each column is a sample, and each row is an OTU. Any object coercible with as.matrix() can be given here, as well as phyloseq, rbiom, SummarizedExperiment, and TreeSummarizedExperiment objects.

weighted

If TRUE, the algorithm takes relative abundances into account. If FALSE, only presence/absence is considered.

pairs

Which combinations of samples should distances be calculated for? The default value (NULL) calculates all-vs-all. Provide a numeric or logical vector specifying positions in the distance matrix to calculate. See examples.

cpus

How many parallel processing threads should be used. The default, n_cpus(), will use all logical CPU cores.

Value

A dist object.

Calculation

In the formulas below, x and y are two columns (samples) from counts. n is the number of rows (OTUs) in counts.

Formula		Example

$$D = \displaystyle \sqrt{\sum_{i = 1}^{n} (x_i - y_i)^{2}}$$

  x <- c(4, 0, 3, 2, 6)
  y <- c(0, 8, 0, 0, 5)
  sqrt(sum((x-y)^2))
  #>  9.69536

References

Legendre P, Caceres M 2013. Beta diversity as the variance of community data: dissimilarity coefficients and partitioning. Ecology Letters. 16(8). https://doi.org/10.1111/ele.12141

See also

Examples

    # Example counts matrix
    ex_counts
#>      A B C  D
#> OTU1 4 0 0  0
#> OTU2 0 8 9 10
#> OTU3 3 0 0  0
#> OTU4 2 0 0  0
#> OTU5 6 5 7  1
    
    # Euclidean weighted distance matrix
    euclidean(ex_counts)
#>           A         B         C
#> B  9.695360                    
#> C 10.535654  2.236068          
#> D 12.409674  4.472136  6.082763
    
    # Euclidean unweighted distance matrix
    euclidean(ex_counts, weighted = FALSE)
#>   A B C
#> B 2    
#> C 2 0  
#> D 2 0 0
    
    # Only calculate distances for A vs all.
    euclidean(ex_counts, pairs = 1:3)
#>          A        B        C
#> B  9.69536                  
#> C 10.53565       NA         
#> D 12.40967       NA       NA