Normalized UniFrac beta diversity metric

Normalized UniFrac beta diversity metric

Usage

normalized_unifrac(counts, tree = NULL, pairs = NULL, cpus = n_cpus())

Arguments

counts

An OTU abundance matrix where each column is a sample, and each row is an OTU. Any object coercible with as.matrix() can be given here, as well as phyloseq, rbiom, SummarizedExperiment, and TreeSummarizedExperiment objects.

tree

A phylo-class object representing the phylogenetic tree for the OTUs in counts. The OTU identifiers given by colnames(counts) must be present in tree. Can be omitted if a tree is embedded with the counts object or as attr(counts, 'tree').

pairs

Which combinations of samples should distances be calculated for? The default value (NULL) calculates all-vs-all. Provide a numeric or logical vector specifying positions in the distance matrix to calculate. See examples.

cpus

How many parallel processing threads should be used. The default, n_cpus(), will use all logical CPU cores.

Calculation

Given \(n\) branches with lengths \(L\) and a pair of samples' abundances (\(A\) and \(B\)) on each of those branches:

$$\displaystyle \frac{\sum_{i = 1}^{n} L_i|\frac{A_i}{A_T} - \frac{B_i}{B_T}|}{\sum_{i = 1}^{n} L_i(\frac{A_i}{A_T} + \frac{B_i}{B_T})}$$

See https://cmmr.github.io/fastbiom/articles/unifrac.html for details and a worked example.

References

Lozupone CA, Hamady M, Kelley ST, Knight R 2007. Quantitative and Qualitative \(\beta\) Diversity Measures Lead to Different Insights into Factors That Structure Microbial Communities. Applied and Environmental Microbiology, 73(5). https://doi.org/10.1128/AEM.01996-06

See also

Examples

    # Example counts matrix
    ex_counts
#>      A B C  D
#> OTU1 4 0 0  0
#> OTU2 0 8 9 10
#> OTU3 3 0 0  0
#> OTU4 2 0 0  0
#> OTU5 6 5 7  1
    
    # UniFrac weighted distance matrix
    normalized_unifrac(ex_counts, tree = ex_tree)
#>            A          B          C
#> B 0.51102941                      
#> C 0.49654112 0.04820717           
#> D 0.69136961 0.22837370 0.27567790
    
    # Only calculate distances for A vs all.
    normalized_unifrac(ex_counts, tree = ex_tree, pairs = 1:3)
#>           A         B         C
#> B 0.5110294                    
#> C 0.4965411        NA          
#> D 0.6913696        NA        NA