Shannon

Shannon alpha diversity metric.

The index considers both the number of different OTUs (richness) and how evenly the observations are distributed among those OTUs (evenness).

Usage

shannon(counts, cpus = n_cpus())

Arguments

counts

An OTU abundance matrix where each column is a sample, and each row is an OTU. Any object coercible with as.matrix() can be given here, as well as phyloseq, rbiom, SummarizedExperiment, and TreeSummarizedExperiment objects.

cpus

How many parallel processing threads should be used. The default, n_cpus(), will use all logical CPU cores.

Value

A numeric vector.

Calculation

Pre-transformation: drop all OTUs with zero abundance.

In the formulas below, \(x\) is a single column (sample) from counts. \(p_i\) is the proportion of the \(i\)-th OTU in the total community.

Formula		Example

$$p_{i} = \displaystyle \frac{x_i}{\sum x}$$ $$D = \displaystyle -\sum_{i = 1}^{n} p_{i}\times\ln(p_{i})$$

  x <- c(4, 0, 3, 2, 6)[-2]
  p <- x / sum(x)
  -sum(p * log(p))
  #>  1.309526

References

Shannon CE, Weaver W 1949. The Mathematical Theory of Communication. University of Illinois Press.

See also

Examples

    # Example counts matrix
    ex_counts
#>      A B C  D
#> OTU1 4 0 0  0
#> OTU2 0 8 9 10
#> OTU3 3 0 0  0
#> OTU4 2 0 0  0
#> OTU5 6 5 7  1
    
    # Shannon diversity values
    shannon(ex_counts)
#>         A         B         C         D 
#> 1.3095258 0.6662784 0.6853142 0.3046361 
    
    # Low diversity
    shannon(c(100, 1, 1, 1, 1)) # 0.22
#> [1] 0.2163426
    
    # High diversity
    shannon(c(20, 20, 20, 20, 20)) # 1.61
#> [1] 1.609438
    
    # Low richness
    shannon(1:3) # 1.01
#> [1] 1.011404
    
    # High richness
    shannon(1:100) # 4.42
#> [1] 4.416898